+++ src/components/BurgerConstructor/BurgerConstructor.module.css
@@ -1,10 +1,8 @@
 .container {
     display: flex;
     flex-direction: column;
-    flex-grow: 1;
-    overflow: auto;
-    composes: custom-scroll from '@ya.praktikum/react-developer-burger-ui-components/dist/ui/common.css'    ;
-    composes: mt-25 pl-4 pr-4 from '@ya.praktikum/react-developer-burger-ui-components/dist/ui/box.css'    ;
+
+    composes: mt-25 pl-4 from '@ya.praktikum/react-developer-burger-ui-components/dist/ui/box.css'    ;
 }
 
 .constructorList{
@@ -12,6 +10,17 @@
     flex-direction: column;
     gap: 16px;
     /* height: 800px; */
+    overflow: auto;
+    flex-grow: 1;
+    composes: custom-scroll from '@ya.praktikum/react-developer-burger-ui-components/dist/ui/common.css'    ;
+    composes:  mb-4 from '@ya.praktikum/react-developer-burger-ui-components/dist/ui/box.css';
+}
+
+.topSide {
+    composes:  mb-4 from '@ya.praktikum/react-developer-burger-ui-components/dist/ui/box.css';
+}
+
+.bottomSide {
     composes:  mb-10 from '@ya.praktikum/react-developer-burger-ui-components/dist/ui/box.css';
 }
 
@@ -26,6 +35,8 @@
 .elementContainer {
     position: relative;
     display: inline-block;
+    composes:  mr-4 from '@ya.praktikum/react-developer-burger-ui-components/dist/ui/box.css';
+
 }
 
 .sumContainer {
@@ -33,6 +44,7 @@
     justify-content: end;
     gap: 40px;
     align-items: center;
+    composes:  mr-4 from '@ya.praktikum/react-developer-burger-ui-components/dist/ui/box.css';
 
 }
