+++ src/components/BurgerIngredients/BurgerIngredients.jsx
@@ -4,6 +4,7 @@ import { Tab, ConstructorElement, CurrencyIcon} from '@ya.praktikum/react-develo
 import styles from './BurgerIngredients.module.css'
 import IngredientDetails from '../IngredientDetails/IngredientDetails'
 import { Counter } from '@ya.praktikum/react-developer-burger-ui-components'
+import ingredientPropTypes from '../../utils/prop-types.js'
 
 const IngredientCard = ({ingredient}) => {
 
@@ -80,38 +81,12 @@ const BurgerIngredients = ({data}) => {
   )
 }
 IngredientCard.propTypes = {
-  ingredient: PropTypes.shape({
-    _id: PropTypes.string.isRequired,
-    name: PropTypes.string.isRequired,
-    type: PropTypes.string.isRequired,
-    proteins: PropTypes.number.isRequired,
-    fat: PropTypes.number.isRequired,
-    carbohydrates: PropTypes.number.isRequired,
-    calories: PropTypes.number.isRequired,
-    price: PropTypes.number.isRequired,
-    image: PropTypes.string.isRequired,
-    image_mobile: PropTypes.string.isRequired,
-    image_large: PropTypes.string.isRequired,
-    __v: PropTypes.number.isRequired,
-  }).isRequired,
+  ingredient: ingredientPropTypes.isRequired,
 };
 
 BurgerIngredients.propTypes = {
   data: PropTypes.arrayOf(
-    PropTypes.shape({
-      _id: PropTypes.string.isRequired,
-      name: PropTypes.string.isRequired,
-      type: PropTypes.string.isRequired,
-      proteins: PropTypes.number.isRequired,
-      fat: PropTypes.number.isRequired,
-      carbohydrates: PropTypes.number.isRequired,
-      calories: PropTypes.number.isRequired,
-      price: PropTypes.number.isRequired,
-      image: PropTypes.string.isRequired,
-      image_mobile: PropTypes.string.isRequired,
-      image_large: PropTypes.string.isRequired,
-      __v: PropTypes.number.isRequired,
-    })
+    ingredientPropTypes
   ).isRequired,
 };
