+++ src/components/IngredientDetails/IngredientDetails.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import styles from './IngredientDetails.module.css'
 import Modal from '../Modal/Modal'
-
+import ingredientPropTypes from '../../utils/prop-types.js'
 
 const Element = ({title, value}) => {
   return(
@@ -40,17 +40,7 @@ Element.propTypes = {
 
 IngredientDetails.propTypes = {
   ingredient: PropTypes.shape({
-    _id: PropTypes.string.isRequired,
-    name: PropTypes.string.isRequired,
-    type: PropTypes.string.isRequired,
-    proteins: PropTypes.number.isRequired,
-    fat: PropTypes.number.isRequired,
-    carbohydrates: PropTypes.number.isRequired,
-    calories: PropTypes.number.isRequired,
-    price: PropTypes.number.isRequired,
-    image: PropTypes.string.isRequired,
-    image_large: PropTypes.string.isRequired,
-    __v: PropTypes.number.isRequired,
+    ingredientPropTypes
   }).isRequired,
   onClose: PropTypes.func.isRequired,
   children: PropTypes.node
