+++ src/components/Modal/Modal.module.css
@@ -1,11 +1,12 @@
 .modal{
-    overflow: hidden;
     position: fixed;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%);
     width: 600px;
-    height: fit-content;
-    z-index: 1000;
+    max-height: 80vh;
+    z-index: 1001;
     margin: auto;
-
     display: flex;
     flex-direction: column;
     align-items: center;
@@ -15,8 +16,6 @@
     background-color: var(--background);
 
     composes: p-10 from '@ya.praktikum/react-developer-burger-ui-components/dist/ui/box.css'    ;
-
-
 }
 
 .header{
