+++ src/components/ModalOverlay/ModalOverlay.jsx
@@ -2,33 +2,17 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import styles from'./ModalOverlay.module.css'
 
-const ModalOverlay = ({ children, onClose }) => {
+const ModalOverlay = ({ onClose }) => {
 
   const handleModalOverlayClick = (e) => {
     if (e.target === e.currentTarget) {
       onClose(e)
     }
   }
-
-  const handleEscPress = (e) => {
-    if (e.key === 'Escape') {
-      onClose(e);
-    }
-  };
-
-  React.useEffect(() => {
-    document.addEventListener('keydown', handleEscPress);
-    return () => document.removeEventListener('keydown', handleEscPress);
-  }, []);
-
-
-  return <div className={styles.modalOverlay} onClick={handleModalOverlayClick}>
-    {children}
-  </div>;
+  return <div className={styles.modalOverlay} onClick={handleModalOverlayClick}/>;
 };
 
 ModalOverlay.propTypes = {
-  children: PropTypes.node.isRequired,
   onClose: PropTypes.func.isRequired,
 };
