+++ src/components/ModalOverlay/ModalOverlay.module.css
@@ -5,12 +5,5 @@
     width: 100%;
     height: 100%;
     z-index: 1000;
-
-    display: flex;
-    justify-content: center;
-    align-items: center;
-
     background-color: rgba(0, 0, 0, 0.6);;
-    overflow:hidden;
-
 }
